fix(rag): search_chunks_hybrid's RRF score is numeric, not double precision
0022 (applied 2026-08-19) fails on every call:

  ERROR: structure of query does not match function result type
  DETAIL: Returned type numeric does not match expected type double
  precision in column 4.

Confirmed live immediately after applying: the RRF expression
(COALESCE(1.0 / (rrf_k + d.rank), 0.0) + COALESCE(1.0 / (rrf_k + s.rank), 0.0))
evaluates to `numeric` (dividing a numeric literal by an integer), but the
function's RETURNS TABLE declares `score float` -- and RETURNS TABLE
requires an exact type match, unlike a plain SELECT which would implicitly
widen numeric to double precision. Cast the fused expression to float
explicitly. Verified the fix as an ad-hoc equivalent SELECT against
production before writing the migration, not just in theory.

0024 is a new corrective migration rather than an edit to 0022's file,
matching the 0018/0020 pattern already established in this branch: 0022
stays historically accurate to what was first applied, 0024 is the fix.
Also fixed in db/schema/chat_db.sql so a fresh recreate from that file
doesn't reintroduce the bug.

No Python-side change -- this is a pure SQL type-casting bug, caught by
actually running the applied function against production rather than only
trusting the migration file. chat_orchestrator's hybrid test suite (8
tests, Python-only, doesn't execute SQL) is unaffected and still green.

Plan: docs/superpowers/plans/2026-08-23-p4-hybrid-agentic-retrieval.md
Phase 1 (post-application fix, found during Phase 2 verification).

+++ db/schema/chat_db.sql
@@ -881,8 +881,10 @@ BEGIN
     ),
     fused AS (
         SELECT COALESCE(d.id, s.id) AS id,
-               COALESCE(1.0 / (rrf_k + d.rank), 0.0)
-             + COALESCE(1.0 / (rrf_k + s.rank), 0.0) AS score
+               (
+                   COALESCE(1.0 / (rrf_k + d.rank), 0.0)
+                 + COALESCE(1.0 / (rrf_k + s.rank), 0.0)
+               )::float AS score
         FROM dense d
         FULL OUTER JOIN sparse s ON s.id = d.id
     )
